+++ desktop/cmp/viewmanager/dialog/SaveAsDialogModel.ts
@@ -6,9 +6,11 @@
  */
 
 import {FormModel} from '@xh/hoist/cmp/form';
-import {HoistModel, managed, XH} from '@xh/hoist/core';
+import {p, strong} from '@xh/hoist/cmp/layout';
+import {HoistModel, managed, SelectOption, XH} from '@xh/hoist/core';
 import {makeObservable, action, observable} from '@xh/hoist/mobx';
 import {ViewManagerModel} from '@xh/hoist/cmp/viewmanager';
+import {some, startCase} from 'lodash';
 
 /**
  * Backing model for ViewManagerModel's SaveAs
@@ -30,15 +32,14 @@ export class SaveAsDialogModel extends HoistModel {
     open() {
         const {parent, formModel} = this,
             src = parent.view,
-            name = parent.ownedViews.some(it => it.name === src.name)
-                ? `Copy of ${src.name}`
-                : src.name;
+            name = some(parent.ownedViews, {name: src.name}) ? `Copy of ${src.name}` : src.name;
 
         formModel.init({
             name,
             group: src.group,
             description: src.description,
-            isShared: false
+            visibility: 'private',
+            isPinned: src.info.isPinned
         });
 
         this.isOpen = true;
@@ -52,10 +53,33 @@ export class SaveAsDialogModel extends HoistModel {
     async saveAsAsync() {
         try {
             await this.doSaveAsAsync().linkTo(this.parent.saveTask);
-            this.close();
         } catch (e) {
             XH.handleException(e);
         }
+    }
+
+    get visibilityOptions(): SelectOption[] {
+        const {parent} = this,
+            ret = [{value: 'private', label: 'Private'}];
+        if (parent.enableSharing) {
+            ret.push({value: 'shared', label: 'Shared'});
+        }
+        if (parent.enableGlobal && parent.manageGlobal) {
+            ret.push({value: 'global', label: startCase(parent.globalDisplayName)});
+        }
+        return ret;
+    }
+
+    get visibilityInfo(): string {
+        switch (this.formModel.values.visibility) {
+            case 'private':
+                return 'Visible to you only.';
+            case 'shared':
+                return 'Visible to ALL users, discoverable via "Shared" tab.';
+            case 'global':
+                return `Visible to ALL users, editable by other ${startCase(this.parent.globalDisplayName)} editors`;
+        }
+        return 'private';
     }
 
     //------------------------
@@ -66,29 +90,60 @@ export class SaveAsDialogModel extends HoistModel {
             fields: [
                 {
                     name: 'name',
-                    rules: [({value}) => this.parent.validateViewNameAsync(value)]
+                    rules: [
+                        ({value}, {visibility}) => {
+                            return this.parent.validateViewNameAsync(
+                                value,
+                                null,
+                                visibility === 'global'
+                            );
+                        }
+                    ]
                 },
                 {name: 'group'},
                 {name: 'description'},
-                {name: 'isShared'}
+                {name: 'isPinned'},
+                {name: 'visibility'}
             ]
         });
     }
 
     private async doSaveAsAsync() {
         let {formModel, parent} = this,
-            {name, group, description, isShared} = formModel.getData(),
-            isValid = await formModel.validateAsync();
+            {name, group, description, visibility, isPinned} = formModel.getData(),
+            isValid = await formModel.validateAsync(),
+            isGlobal = visibility == 'global',
+            isShared = visibility == 'shared';
 
         if (!isValid) return;
 
-        return parent.saveAsAsync({
+        if (isGlobal) {
+            const message = [
+                p(
+                    `This ${parent.typeDisplayName} will become a ${parent.globalDisplayName} ${parent.typeDisplayName} visible to all other ${XH.appName} users.`
+                ),
+                p(strong('Are you sure you want to proceed?'))
+            ];
+            const confirmed = await XH.confirm({
+                message,
+                confirmProps: {
+                    outlined: true,
+                    autoFocus: false,
+                    intent: 'primary'
+                }
+            });
+            if (!confirmed) return;
+        }
+
+        await parent.saveAsAsync({
             name: name.trim(),
             group: group?.trim(),
             description: description?.trim(),
-            isPinned: true,
+            isPinned,
+            isGlobal,
             isShared,
             value: parent.getValue()
         });
+        this.close();
     }
 }